[#85] Reorganize routes into (auth) dashboard and (olds) legacy groups
