+++ frontend/src/app/(auth)/sop-library/page.js
@@ -1,0 +1,16 @@
+"use client";
+
+import { FeedbackSection } from "@/components";
+
+const SOPLibraryPage = () => {
+  return (
+    <div>
+      <h1>SOP Library</h1>
+      <div className="py-8">
+        <FeedbackSection />
+      </div>
+    </div>
+  );
+};
+
+export default SOPLibraryPage;
