+++ frontend/src/app/(auth)/validations/page.js
@@ -1,0 +1,16 @@
+"use client";
+
+import { FeedbackSection } from "@/components";
+
+const ValidationsPage = () => {
+  return (
+    <div>
+      <h1>My validations</h1>
+      <div className="py-8">
+        <FeedbackSection />
+      </div>
+    </div>
+  );
+};
+
+export default ValidationsPage;
