+++ frontend/src/app/(olds)/publications/[id]/page.js
