+++ frontend/src/app/(olds)/publications/[id]/validation/page.js
