+++ frontend/src/app/(olds)/publications/layout.js
@@ -1,14 +1,11 @@
 import { UserContextProvider } from "@/context";
 import { auth } from "@/lib";
-import { DashboardLayout } from "@/components";
 
 const PublicationLayout = async ({ children }) => {
   const session = await auth.getSession();
   const abilities = session?.abilities || [];
   return (
-    <UserContextProvider abilities={abilities}>
-      <DashboardLayout user={session}>{children}</DashboardLayout>
-    </UserContextProvider>
+    <UserContextProvider abilities={abilities}>{children}</UserContextProvider>
   );
 };
